move validation to parser package
